# chore: add internal vite config