--- a/.prettierignore
+++ b/.prettierignore
@@ -6,3 +6,5 @@ dist
 vite.config.ts.timestamp-*
 
 packages/tuono-lazy-fn-vite-plugin/tests/sources/*
+
+e2e/fixtures/*/test-results