--- a/devtools/vite-config/package.json
+++ b/devtools/vite-config/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "vite-config",
+  "version": "1.0.0",
+  "type": "module",
+  "private": true,
+  "exports": {
+    "types": "./dist/index.d.ts",
+    "default": "./dist/index.js"
+  },
+  "scripts": {
+    "build": "tsc"
+  },
+  "devDependencies": {
+    "rollup-plugin-preserve-directives": "0.4.0",
+    "vite": "5.4.14",
+    "vite-plugin-externalize-deps": "0.9.0",
+    "vite-tsconfig-paths": "5.1.4"
+  }
+}