--- a/devtools/vite-config/package.json
+++ b/devtools/vite-config/package.json
@@ -13,6 +13,7 @@
   "devDependencies": {
     "rollup-plugin-preserve-directives": "0.4.0",
     "vite": "5.4.14",
+    "vite-plugin-dts": "4.5.0",
     "vite-plugin-externalize-deps": "0.9.0",
     "vite-tsconfig-paths": "5.1.4"
   }