--- a/devtools/vite-config/package.json
+++ b/devtools/vite-config/package.json
@@ -11,7 +11,9 @@
     "build": "tsc"
   },
   "devDependencies": {
+    "oxc-transform": "0.50.0",
     "rollup-plugin-preserve-directives": "0.4.0",
+    "unplugin-isolated-decl": "0.11.0",
     "vite": "5.4.14",
     "vite-plugin-dts": "4.5.0",
     "vite-plugin-externalize-deps": "0.9.0",