--- a/devtools/vite-config/package.json
+++ b/devtools/vite-config/package.json
@@ -15,8 +15,6 @@
     "rollup-plugin-preserve-directives": "0.4.0",
     "unplugin-isolated-decl": "0.11.0",
     "vite": "5.4.14",
-    "vite-plugin-dts": "4.5.0",
-    "vite-plugin-externalize-deps": "0.9.0",
-    "vite-tsconfig-paths": "5.1.4"
+    "vite-plugin-externalize-deps": "0.9.0"
   }
 }