--- a/devtools/vite-config/src/index.ts
+++ b/devtools/vite-config/src/index.ts
@@ -2,16 +2,21 @@ import type { UserConfig } from 'vite'
 import { preserveDirectives } from 'rollup-plugin-preserve-directives'
 import { externalizeDeps } from 'vite-plugin-externalize-deps'
 import tsconfigPaths from 'vite-tsconfig-paths'
+import dts from 'vite-plugin-dts'
 
 interface Options {
   /** Entry file, e.g. `./src/index.ts` */
   entry: string | Array<string>
+
+  /** Source directory used for type generation, e.g. `./src` */
+  srcDir: string
 }
 
 export function defineViteConfig(options: Options): UserConfig {
+  const outDir = 'dist'
   return {
     build: {
-      outDir: 'dist',
+      outDir,
       minify: false,
       sourcemap: true,
       lib: {
@@ -25,6 +30,15 @@ export function defineViteConfig(options: Options): UserConfig {
         },
       },
     },
-    plugins: [externalizeDeps(), preserveDirectives(), tsconfigPaths()],
+    plugins: [
+      externalizeDeps(),
+      preserveDirectives(),
+      tsconfigPaths(),
+      dts({
+        outDir: `${outDir}/esm`,
+        entryRoot: options.srcDir,
+        include: options.srcDir,
+      }),
+    ],
   }
 }