--- a/devtools/vite-config/src/index.ts
+++ b/devtools/vite-config/src/index.ts
@@ -2,31 +2,31 @@ import type { UserConfig } from 'vite'
 import { preserveDirectives } from 'rollup-plugin-preserve-directives'
 import { externalizeDeps } from 'vite-plugin-externalize-deps'
 import tsconfigPaths from 'vite-tsconfig-paths'
-import dts from 'vite-plugin-dts'
+import UnpluginIsolatedDecl from 'unplugin-isolated-decl/vite'
 
 interface Options {
   /** Entry file, e.g. `./src/index.ts` */
   entry: string | Array<string>
-
-  /** Source directory used for type generation, e.g. `./src` */
-  srcDir: string
 }
 
 export function defineViteConfig(options: Options): UserConfig {
+  const { entry } = options
   const outDir = 'dist'
+
   return {
     build: {
       outDir,
       minify: false,
       sourcemap: true,
       lib: {
-        entry: options.entry,
+        entry,
         formats: ['es'],
         fileName: 'esm/[name]',
       },
       rollupOptions: {
         output: {
           preserveModules: true,
+          entryFileNames: 'esm/[name].js',
         },
       },
     },
@@ -34,11 +34,12 @@ export function defineViteConfig(options: Options): UserConfig {
       externalizeDeps(),
       preserveDirectives(),
       tsconfigPaths(),
-      dts({
-        outDir: `${outDir}/esm`,
-        entryRoot: options.srcDir,
-        include: options.srcDir,
-      }),
+      UnpluginIsolatedDecl({ transformer: 'oxc' }),
+      // dts({
+      //   outDir: `${outDir}/esm`,
+      //   entryRoot: options.srcDir,
+      //   include: options.srcDir,
+      // }),
     ],
   }
 }