--- a/devtools/vite-config/tsconfig.json
+++ b/devtools/vite-config/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "include": ["src"],
+  "compilerOptions": {
+    "moduleResolution": "node16",
+    "module": "Node16",
+    "declaration": true,
+    "outDir": "dist",
+    "strict": true
+  }
+}