--- a/package.json
+++ b/package.json
@@ -21,9 +21,9 @@
   "license": "MIT",
   "devDependencies": {
     "@eslint/js": "9.20.0",
+    "@playwright/test": "1.50.1",
     "@types/node": "22.13.1",
     "eslint": "9.20.0",
-    "@playwright/test": "^1.50.1",
     "eslint-import-resolver-typescript": "3.7.0",
     "eslint-plugin-import": "2.31.0",
     "eslint-plugin-react": "7.37.4",