--- a/packages/tuono-fs-router-vite-plugin/package.json
+++ b/packages/tuono-fs-router-vite-plugin/package.json
@@ -44,7 +44,6 @@
     "vite": "^5.4.14"
   },
   "devDependencies": {
-    "@tanstack/config": "0.7.13",
     "@types/babel__core": "7.20.5",
     "vite-config": "workspace:*",
     "vitest": "2.1.9"