--- a/packages/tuono-fs-router-vite-plugin/src/index.ts
+++ b/packages/tuono-fs-router-vite-plugin/src/index.ts
@@ -1,4 +1,4 @@
-import { normalize } from 'path'
+import { normalize } from 'node:path'
 
 import type { Plugin } from 'vite'