--- a/packages/tuono-fs-router-vite-plugin/vite.config.ts
+++ b/packages/tuono-fs-router-vite-plugin/vite.config.ts
@@ -2,5 +2,4 @@ import { defineViteConfig } from 'vite-config'
 
 export default defineViteConfig({
   entry: './src/index.ts',
-  srcDir: './src',
 })