--- a/packages/tuono-router/package.json
+++ b/packages/tuono-router/package.json
@@ -32,14 +32,8 @@
   ],
   "exports": {
     ".": {
-      "import": {
-        "types": "./dist/esm/index.d.ts",
-        "default": "./dist/esm/index.js"
-      },
-      "require": {
-        "types": "./dist/cjs/index.d.cts",
-        "default": "./dist/cjs/index.cjs"
-      }
+      "types": "./dist/esm/index.d.ts",
+      "default": "./dist/esm/index.js"
     },
     "./package.json": "./package.json"
   },
@@ -50,13 +44,13 @@
     "react-intersection-observer": "^9.13.0"
   },
   "devDependencies": {
-    "@tanstack/config": "0.7.13",
     "@testing-library/react": "16.2.0",
     "@types/react": "19.0.8",
     "@vitejs/plugin-react-swc": "3.7.2",
     "happy-dom": "16.8.1",
     "react": "19.0.0",
     "vite": "5.4.14",
+    "vite-config": "workspace:*",
     "vitest": "2.1.9"
   }
 }