--- a/packages/tuono-router/vite.config.ts
+++ b/packages/tuono-router/vite.config.ts
@@ -1,21 +1,19 @@
 /// <reference types="vitest" />
 /// <reference types="vite/client" />
 import { defineConfig, mergeConfig } from 'vitest/config'
-import { tanstackBuildConfig } from '@tanstack/config/build'
+import { defineViteConfig } from 'vite-config'
 import react from '@vitejs/plugin-react-swc'
 
-const config = defineConfig({
-  plugins: [react()],
-  test: {
-    name: 'tuono-router',
-    environment: 'happy-dom',
-    globals: true,
-  },
-})
-
 export default mergeConfig(
-  config,
-  tanstackBuildConfig({
+  defineConfig({
+    plugins: [react()],
+    test: {
+      name: 'tuono-router',
+      environment: 'happy-dom',
+      globals: true,
+    },
+  }),
+  defineViteConfig({
     entry: './src/index.ts',
     srcDir: './src',
   }),