--- a/packages/tuono-router/vite.config.ts
+++ b/packages/tuono-router/vite.config.ts
@@ -15,6 +15,5 @@ export default mergeConfig(
   }),
   defineViteConfig({
     entry: './src/index.ts',
-    srcDir: './src',
   }),
 )