--- a/packages/tuono-router/vite.config.ts
+++ b/packages/tuono-router/vite.config.ts
@@ -5,6 +5,9 @@ import { defineViteConfig } from 'vite-config'
 import react from '@vitejs/plugin-react-swc'
 
 export default mergeConfig(
+  defineViteConfig({
+    entry: './src/index.ts',
+  }),
   defineConfig({
     plugins: [react()],
     test: {
@@ -12,8 +15,5 @@ export default mergeConfig(
       environment: 'happy-dom',
       globals: true,
     },
-  }),
-  defineViteConfig({
-    entry: './src/index.ts',
   }),
 )