--- a/packages/tuono/package.json
+++ b/packages/tuono/package.json
@@ -113,7 +113,8 @@
     "@types/react-dom": "19.0.3",
     "react": "19.0.0",
     "react-dom": "19.0.0",
-    "vitest": "2.1.9"
+    "vitest": "2.1.9",
+    "vite-config": "workspace:*"
   },
   "sideEffects": false,
   "keywords": [