--- a/packages/tuono/package.json
+++ b/packages/tuono/package.json
@@ -105,7 +105,6 @@
     "web-streams-polyfill": "^4.0.0"
   },
   "devDependencies": {
-    "@tanstack/config": "0.7.13",
     "@types/babel__core": "7.20.5",
     "@types/babel__traverse": "7.20.6",
     "@types/node": "22.13.1",