--- a/packages/tuono/package.json
+++ b/packages/tuono/package.json
@@ -7,9 +7,9 @@
     "dev": "vite build --watch",
     "build": "vite build",
     "lint": "eslint .",
-    "format": "prettier --write --ignore-unknown --ignore-path ../../.prettierignore .",
-    "format:check": "prettier --check --ignore-unknown --ignore-path ../../.prettierignore .",
-    "types": "tsc --noEmit",
+    "format": "prettier --check --ignore-unknown --ignore-path ../../.prettierignore .",
+    "format:fix": "prettier --write --ignore-unknown --ignore-path ../../.prettierignore .",
+    "typecheck": "tsc --noEmit",
     "test:watch": "vitest",
     "test": "vitest run"
   },