--- a/packages/tuono/vite.config.ts
+++ b/packages/tuono/vite.config.ts
@@ -18,6 +18,5 @@ export default mergeConfig(
       './src/ssr/index.ts',
       './src/hydration/index.tsx',
     ],
-    srcDir: './src',
   }),
 )