--- a/pnpm-workspace.yaml
+++ b/pnpm-workspace.yaml
@@ -1,4 +1,5 @@
 packages:
+  - 'devtools/*'
   - 'packages/*'
   - 'examples/*'
   - 'e2e/fixtures/*'