<!--
👋 Thank you for your Pull Request 🙏
-->

### Checklist

- [x] I have read [Contributing > Pull requests](https://tuono.dev/documentation/contributing/pull-requests)

### Related issue

Fixes #566

<!-- Replace the content with "None" if you haven't an issue to link -->

### Overview

- Added new devtools folder to house all packages that will be used internally
  - In the future this could house shared config between different packages (eslint config, playwright setup, ...)
- Added `vite-config` package which expose a `defineViteConfig` function that can be used to define `vite` config 
  (who would have guess that?)
- Removed `@tanstack/config`
- Removed cjs build and related exports.
- Streamlined package script names

---

### TODOs

- [x] e2e tests are working
- [x] add lint and format check

<!--

Explain the context and why you're making that change.
What is the problem you're trying to solve?
If a new feature is being added,
describe the intended use case that feature fulfills.

-->