@marcalexiei 2025-02-16
### Simple performance compare

| Branch  | GitHub Actions | MacBook Air M3 |
|--------|--------|--------|
| `main` | [~12s](https://github.com/tuono-labs/tuono/actions/runs/13346532122/job/37277630904) | ~5.4s |
| `this PR` | [~4s](https://github.com/tuono-labs/tuono/actions/runs/13355854131/job/37298504754) | ~2.2s |